WIP: refactoring due to KnowledgeBase changes
